applications/dash-multi-page-penguins/using-dash-pages/requirements.txt
dash
pandas
dash-bootstrap-components
rsconnect-python
python-dotenv